Provide reason for skipping bignum tests
It's good for the user to know why tests are being skipped, hence it's
implemented here.  The nicest way to do this was to use the `skip_all`
option to `plan` of `Test::More`, hence `Test::More` is used instead of
`Test`.  This shouldn't be a problem, since `Test::More` has been in core
for a long time now.